fix: Rename logging folder to avoid naming conflicts (and circular import issues)
