+++ archivebox/cli/archivebox_status.py
@@ -10,7 +10,7 @@ from typing import Optional, List, IO
 
 from ..main import status, docstring
 from ..config import OUTPUT_DIR
-from ..logging import SmartFormatter, reject_stdin
+from ..logging_util import SmartFormatter, reject_stdin
 
 
 @docstring(status.__doc__)
