+++ archivebox/extractors/git.py
@@ -22,7 +22,7 @@ from ..config import (
     GIT_DOMAINS,
     CHECK_SSL_VALIDITY
 )
-from ..logging import TimedProgress
+from ..logging_util import TimedProgress
