+++ archivebox/extractors/title.py
@@ -18,7 +18,7 @@ from ..config import (
     CURL_VERSION,
     CURL_USER_AGENT,
 )
-from ..logging import TimedProgress
+from ..logging_util import TimedProgress
 
 
 HTML_TITLE_REGEX = re.compile(
